Move `Modal(Impl)` to `modals` package
